Completed Core Code
The robot scans along the middlemost row of the image. If it detects a white pixel, it adds wherever it found it to a variable. It also stores how many white pixels it has come across so far.
It gets the average position of the white pixel. From that, it calculates the distance between the average area of white pixels and the centre of the image. If that number is negative, it means that the white pixels were found in an area to the right of the centre. If that number is positive, it means that the white pixels were found in an area to the left of the centre. If the number is 0, then the area where the white pixels were found in the centre.
Depending on whether the number is negative or positive, it makes one wheel slightly faster than the other wheel. This process repeats every frame.
If the speed is changed, the robot may not be able to make sharp corners. We found 13 was a decent value to have one wheel, while the other stays at 10.

The end result is a slight swaying motion as the robot always attempts to get the line in the exact centre. It still completes the maze, it's just a bit drunk.