AVC Completion Code
The robot can now travel through intersections and sharp corners.

It detects is there there is white in the top row, or the two ends of the camera view. If it's just in the top, it goes forward, exactly the same as to how it worked in the core code. If it detects white directly to the left, it will travel to the left. If it detects white to the right, it will turn right. At intersections, where white is detected to the left and the right, it will turn left.
If the robot comes to a dead end, it will end up spinning in place. If there is no more white, but black is detected, it assumes that is the flag. This is a poor method of flag detection, as it may be possible for the camera to go off-screen when it spins, in which case, it will assume it has reached the flag and stop.
The code is also now split into functions. There are functions for detecting the white lines and moving the robot.