robot-challenge.cpp
This version of the robot is built to navigate the Challenge maze. It works on the "left wall principle" - that you can reach the end of most mazes by following the left wall. The robot detects red on it's left and follows it, turning right when it would run into a wall. To account for the distance between the camera view and the robot, there's a delay between the robot knowing it needs to turn and the robot actually turning. If there's no wall to follow on the left, the robot will turn left instead of right. However, the robot needed very specific numbers for how long to delay before turning and such - it works on the given Challenge maze, but other mazes with thicker/thinner walls or wider hallways wouldn't be navigable to the robot in its current state.